Fix adds exception and increase use of types

+++ src/books/books.service.ts
@@ -1,7 +1,7 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Book } from './book.entity';
-import {Repository} from 'typeorm'
+import { Repository } from 'typeorm';
 import { CreateBookDto } from './dtos/create-book.dto';
 
 @Injectable()
@@ -13,24 +13,30 @@ export class BooksService {
     return this.repo.save(book);
   }
 
-  async findAll() {
-    const allbooks = await this.repo.find();
+  async findAll(): Promise<Book[]> {
+    const allbooks: Book[] = await this.repo.find();
+
     return allbooks;
   }
-  async findOne(id: number) {
-    const book = await this.repo.findOne({ where: { id: id } });
+  async findOne(id: number): Promise<Book> {
+    const book: Book = await this.repo.findOne({ where: { id: id } });
     return book;
   }
 
-  async update(id: number, attrs: Partial<Book>) {
-    let book = await this.repo.findOne({ where: { id: id } });
+  async update(id: number, attrs: Partial<Book>): Promise<Book> {
+    let book: Book = await this.repo.findOne({ where: { id: id } });
 
     book = Object.assign(book, attrs);
     return this.repo.save(book);
   }
 
-  async remove(id: number) {
-    let book = await this.repo.findOne({ where: { id: id } });
-    this.repo.remove(book);
+  async remove(id: number): Promise<void> {
+    const book: Book = await this.repo.findOne({ where: { id: id } });
+
+    if (!book) {
+      throw new NotFoundException(`Book not found`);
+    } else {
+      await this.repo.remove(book);
+    }
   }
 }
